chore(conv-creation): Move files to mvvm sub-packages
* close dialog upon saving
* colorize title lines
* simplyfy preview code fragment
