+++ app/src/main/java/com/nextcloud/talk/utils/preview/ComposePreviewUtils.kt
@@ -27,8 +27,8 @@ import com.nextcloud.talk.chat.viewmodels.ChatViewModel
 import com.nextcloud.talk.contacts.ContactsRepository
 import com.nextcloud.talk.contacts.ContactsRepositoryImpl
 import com.nextcloud.talk.contacts.ContactsViewModel
-import com.nextcloud.talk.conversationcreation.ConversationCreationRepositoryImpl
-import com.nextcloud.talk.conversationcreation.ConversationCreationViewModel
+import com.nextcloud.talk.conversationcreation.data.ConversationCreationRepositoryImpl
+import com.nextcloud.talk.conversationcreation.viewmodel.ConversationCreationViewModel
 import com.nextcloud.talk.conversationlist.data.OfflineConversationsRepository
 import com.nextcloud.talk.conversationlist.data.network.ConversationsNetworkDataSource
 import com.nextcloud.talk.conversationlist.data.network.OfflineFirstConversationsRepository
